Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I tried to figure out why this fails, but didn't have any luck.

I thought this was going to be based on the mistral-lib spec? It looks like this is very different and doesn't have any of the python3 support. I'm not sure if that is required now, but I think it should be.

https://github.com/rdo-packages/mistral-lib-distgit/blob/rpm-master/python-mistral-lib.spec

Patch-set: 1
Label: Code-Review=-1